moving make_store_config features back to sled, fxies #562
