+++ crates/matrix-sdk/src/store.rs
@@ -29,62 +29,7 @@
 //! [`StoreConfig`]: crate::config::StoreConfig
 //! [`ClientBuilder::store_config()`]: crate::ClientBuilder::store_config
 
-#[cfg(any(feature = "indexeddb-state-store", feature = "indexeddb-crypto-store"))]
+#[cfg(feature = "indexeddb")]
 pub use matrix_sdk_indexeddb::*;
-#[cfg(any(feature = "sled-state-store", feature = "sled-crypto-store"))]
+#[cfg(feature = "sled")]
 pub use matrix_sdk_sled::*;
-
-// FIXME Move these two methods back to the matrix-sdk-sled crate once weak
-// dependency features are stable and we decide to bump the MSRV.
-
-/// Create a [`StoreConfig`] with an opened sled [`StateStore`] that uses the
-/// given path and passphrase. If `encryption` is enabled, a [`CryptoStore`]
-/// with the same parameters is also opened.
-///
-/// [`StoreConfig`]: #crate::config::StoreConfig
-#[cfg(any(feature = "sled-state-store", feature = "sled-crypto-store"))]
-pub fn make_store_config(
-    path: impl AsRef<std::path::Path>,
-    passphrase: Option<&str>,
-) -> Result<crate::config::StoreConfig, OpenStoreError> {
-    #[cfg(all(feature = "e2e-encryption", feature = "sled-state-store"))]
-    {
-        let (state_store, crypto_store) = open_stores_with_path(path, passphrase)?;
-        Ok(crate::config::StoreConfig::new().state_store(state_store).crypto_store(crypto_store))
-    }
-
-    #[cfg(all(feature = "e2e-encryption", not(feature = "sled-state-store")))]
-    {
-        let crypto_store = CryptoStore::open_with_passphrase(path, passphrase)?;
-        Ok(crate::config::StoreConfig::new().crypto_store(Box::new(crypto_store)))
-    }
-
-    #[cfg(not(feature = "e2e-encryption"))]
-    {
-        let state_store = if let Some(passphrase) = passphrase {
-            StateStore::open_with_passphrase(path, passphrase)?
-        } else {
-            StateStore::open_with_path(path)?
-        };
-
-        Ok(crate::config::StoreConfig::new().state_store(Box::new(state_store)))
-    }
-}
-
-/// Create a [`StateStore`] and a [`CryptoStore`] that use the same database and
-/// passphrase.
-#[cfg(all(feature = "sled-state-store", feature = "sled-crypto-store"))]
-fn open_stores_with_path(
-    path: impl AsRef<std::path::Path>,
-    passphrase: Option<&str>,
-) -> Result<(Box<StateStore>, Box<CryptoStore>), OpenStoreError> {
-    if let Some(passphrase) = passphrase {
-        let state_store = StateStore::open_with_passphrase(path, passphrase)?;
-        let crypto_store = state_store.open_crypto_store(Some(passphrase))?;
-        Ok((Box::new(state_store), Box::new(crypto_store)))
-    } else {
-        let state_store = StateStore::open_with_path(path)?;
-        let crypto_store = state_store.open_crypto_store(None)?;
-        Ok((Box::new(state_store), Box::new(crypto_store)))
-    }
-}
